Adds support for redirects
This helps us create alternate URLs for users to type
and remember.

endoflife.date/k8s is much easier to use and type, for eg

The _redirects file renders content from the alternate_urls key
in every page on the site. This lets us create alternate urls for
other pages as well (not just tools/).